+++ Library/Library/Executable.cpp
@@ -86,7 +86,7 @@ WCHAR* Executable::copy(TCchar* src, WCHAR* vector, int nChars) {
 int i;
 int n;
 
-  if (!src) return 0;   n = _tcslen(src);   if (n >= nChars) return 0;
+  if (!src) return 0;   n = (int) _tcslen(src);   if (n >= nChars) return 0;
 
   for (i = 0; i < n; i++) vector[i] = *src++;   vector[i] = 0;
